+++ lib/src/model/country.dart
@@ -1,264 +1,261 @@
 // Copyright (c) 2019-2021, Ben Hills. Use of this source code is governed by a
 // MIT license that can be found in the LICENSE file.
 
-/// A class that provides a list of the countries supported by the iTunes API.
-class Country {
-  final String _countryCode;
+enum Country {
+  none(code: ''),
+  afghanistan(code: 'af'),
+  alandislands(code: 'ax'),
+  albania(code: 'al'),
+  algeria(code: 'dz'),
+  americansamoa(code: 'as'),
+  andorra(code: 'ad'),
+  angolia(code: 'ao'),
+  anguilla(code: 'ai'),
+  antarctica(code: 'aq'),
+  antiguaandbarbuda(code: 'ag'),
+  argentina(code: 'ar'),
+  armenia(code: 'am'),
+  aruba(code: 'aw'),
+  australia(code: 'au'),
+  austria(code: 'at'),
+  azerbaijan(code: 'az'),
+  bahamas(code: 'bs'),
+  bahrain(code: 'bh'),
+  bangladesh(code: 'bd'),
+  barbados(code: 'bb'),
+  belarus(code: 'by'),
+  belgium(code: 'be'),
+  belize(code: 'bz'),
+  benin(code: 'bj'),
+  bermuda(code: 'bm'),
+  bhutan(code: 'bt'),
+  bolivia(code: 'bo'),
+  bonaire(code: 'bq'),
+  bosniaAndHerzegovina(code: 'ba'),
+  botswana(code: 'bw'),
+  bouvetisland(code: 'bv'),
+  brazil(code: 'br'),
+  britishIndianOceanTerrirory(code: 'io'),
+  britishVirginIslands(code: 'vg'),
+  bruneiDarussalam(code: 'bn'),
+  bulgaria(code: 'bg'),
+  burkinafaso(code: 'bf'),
+  burundi(code: 'bi'),
+  caboverde(code: 'cv'),
+  cambodia(code: 'kh'),
+  cameroon(code: 'cm'),
+  canada(code: 'ca'),
+  caymanIslands(code: 'ky'),
+  centralAfricanRepublic(code: 'cf'),
+  chad(code: 'td'),
+  chile(code: 'cl'),
+  china(code: 'cn'),
+  christmasIsland(code: 'cx'),
+  cocosIslands(code: 'cc'),
+  colombia(code: 'co'),
+  comoros(code: 'km'),
+  congo(code: 'cg'),
+  congoDemocraticRepublicof(code: 'cd'),
+  cookIslands(code: 'ck'),
+  costarica(code: 'cr'),
+  coteDivoire(code: 'ci'),
+  croatia(code: 'hr'),
+  cuba(code: 'cu'),
+  curacao(code: 'cw'),
+  cyprus(code: 'cy'),
+  czechia(code: 'cz'),
+  denmark(code: 'dk'),
+  djibouti(code: 'dj'),
+  dominica(code: 'dm'),
+  dominicanRepublic(code: 'do'),
+  ecuador(code: 'ec'),
+  egypt(code: 'eg'),
+  elSalvador(code: 'sv'),
+  equatorialGuinea(code: 'gq'),
+  eritrea(code: 'er'),
+  estonia(code: 'ee'),
+  ethiopia(code: 'et'),
+  falklandIslands(code: 'fk'),
+  faroeIslands(code: 'fo'),
+  fiji(code: 'fj'),
+  finland(code: 'fi'),
+  france(code: 'fr'),
+  frenchGuiana(code: 'gf'),
+  frenchPolynesia(code: 'pf'),
+  frenchSouthernTerritories(code: 'tf'),
+  gabon(code: 'ga'),
+  gambia(code: 'gm'),
+  georgia(code: 'ge'),
+  germany(code: 'de'),
+  ghana(code: 'gh'),
+  gibraltar(code: 'gi'),
+  greece(code: 'gr'),
+  greenland(code: 'gl'),
+  grenada(code: 'gd'),
+  guadeloupe(code: 'gp'),
+  guam(code: 'gu'),
+  guatemala(code: 'gt'),
+  guernsey(code: 'gg'),
+  guinea(code: 'gn'),
+  guinea_bissau(code: 'gw'),
+  guyana(code: 'gy'),
+  haiti(code: 'ht'),
+  heardIslandAnd_McDonaldIslands(code: 'hm'),
+  honduras(code: 'hn'),
+  hongKong(code: 'hk'),
+  hungary(code: 'hu'),
+  iceland(code: 'is'),
+  india(code: 'in'),
+  indonesia(code: 'id'),
+  iran(code: 'ir'),
+  iraq(code: 'iq'),
+  ireland(code: 'ie'),
+  isleOfMan(code: 'im'),
+  israel(code: 'il'),
+  italy(code: 'it'),
+  jamaica(code: 'jm'),
+  japan(code: 'jp'),
+  jersey(code: 'je'),
+  jordan(code: 'jo'),
+  kazakhstan(code: 'kz'),
+  kenya(code: 'ke'),
+  kiribati(code: 'ki'),
+  kuwait(code: 'kw'),
+  kyrgyzstan(code: 'kg'),
+  laos(code: 'la'),
+  latvia(code: 'lv'),
+  lebanon(code: 'lb'),
+  lesotho(code: 'ls'),
+  liberia(code: 'lr'),
+  libya(code: 'ly'),
+  liechtenstein(code: 'li'),
+  lithuania(code: 'lt'),
+  luxembourg(code: 'lu'),
+  macao(code: 'mo'),
+  macedonia(code: 'mk'),
+  madagascar(code: 'mg'),
+  malawi(code: 'mw'),
+  malaysia(code: 'my'),
+  maldives(code: 'mv'),
+  mali(code: 'ml'),
+  malta(code: 'mt'),
+  marshallIslands(code: 'mh'),
+  martinique(code: 'mq'),
+  mauritania(code: 'mr'),
+  mauritius(code: 'mu'),
+  mayotte(code: 'yt'),
+  mexico(code: 'mx'),
+  micronesia(code: 'fm'),
+  moldova(code: 'md'),
+  monaco(code: 'mc'),
+  mongolia(code: 'mn'),
+  montenegro(code: 'me'),
+  montserrat(code: 'ms'),
+  morocco(code: 'ma'),
+  mozambique(code: 'mz'),
+  myanmar(code: 'mm'),
+  namibia(code: 'na'),
+  nauru(code: 'nr'),
+  nepal(code: 'np'),
+  netherlands(code: 'nl'),
+  newCaledonia(code: 'nc'),
+  newZealand(code: 'nz'),
+  nicaragua(code: 'ni'),
+  niger(code: 'ne'),
+  nigeria(code: 'ng'),
+  niue(code: 'nu'),
+  norfolkIsland(code: 'nf'),
+  northKorea(code: 'kp'),
+  northernMarianaIslands(code: 'mp'),
+  norway(code: 'no'),
+  oman(code: 'om'),
+  pakistan(code: 'pk'),
+  palau(code: 'pw'),
+  palestine(code: 'ps'),
+  panama(code: 'pa'),
+  papuaNewGuinea(code: 'pg'),
+  paraguay(code: 'py'),
+  peru(code: 'pe'),
+  philippines(code: 'ph'),
+  pitcairn(code: 'pn'),
+  poland(code: 'pl'),
+  portugal(code: 'pt'),
+  puerto_rico(code: 'pr'),
+  qatar(code: 'qa'),
+  reunion(code: 're'),
+  romania(code: 'ro'),
+  russianFederation(code: 'ru'),
+  rwanda(code: 'rw'),
+  saintBarthelemy(code: 'bl'),
+  saintHelena(code: 'sh'),
+  saintKittsAndNevis(code: 'kn'),
+  saintLucia(code: 'lc'),
+  saintMartin(code: 'mf'),
+  saintPierreAndMiquelon(code: 'pm'),
+  saintVincentAndTheGrenadines(code: 'vc'),
+  samoa(code: 'ws'),
+  sanMarino(code: 'sm'),
+  saoTomeAndPrincipe(code: 'st'),
+  saudiArabia(code: 'sa'),
+  senegal(code: 'sn'),
+  serbia(code: 'rs'),
+  seychelles(code: 'sc'),
+  sierraLeone(code: 'sl'),
+  singapore(code: 'sg'),
+  sintMaarten(code: 'sx'),
+  slovakia(code: 'sk'),
+  slovenia(code: 'si'),
+  solomonIslands(code: 'sb'),
+  somalia(code: 'so'),
+  southAfrica(code: 'za'),
+  southGeorgiaAndTheSouthSandwichIslands(code: 'gs'),
+  southKorea(code: 'kr'),
+  southSudan(code: 'ss'),
+  spain(code: 'es'),
+  sriLanka(code: 'lk'),
+  sudan(code: 'sd'),
+  suriname(code: 'sr'),
+  svalbardAndJanMayen(code: 'sj'),
+  swaziland(code: 'sz'),
+  sweden(code: 'se'),
+  switzerland(code: 'ch'),
+  syrianArabRepublic(code: 'sy'),
+  taiwan(code: 'tw'),
+  tajikistan(code: 'tj'),
+  tanzania(code: 'tz'),
+  thailand(code: 'th'),
+  timorLeste(code: 'tl'),
+  togo(code: 'tg'),
+  tokelau(code: 'tk'),
+  tonga(code: 'to'),
+  trinidadAndTobago(code: 'tt'),
+  tunisia(code: 'tn'),
+  turkey(code: 'tr'),
+  turkmenistan(code: 'tm'),
+  turksAndCaicosIslands(code: 'tc'),
+  tuvalu(code: 'tv'),
+  uganda(code: 'ug'),
+  ukraine(code: 'ua'),
+  unitedArabEmirates(code: 'ae'),
+  unitedKingdom(code: 'gb'),
+  unitedStates(code: 'us'),
+  unitedStatesMinorOutlyingIslands(code: 'um'),
+  uruguay(code: 'uy'),
+  usVirginIslands(code: 'vi'),
+  uzbekistan(code: 'uz'),
+  vanuatu(code: 'vu'),
+  vaticanCity(code: 'va'),
+  venezuela(code: 've'),
+  vietNam(code: 'vn'),
+  wallisAndFutuna(code: 'wf'),
+  westernSahara(code: 'eh'),
+  yemen(code: 'ye'),
+  zambia(code: 'zm'),
+  zimbabwe(code: 'zw');
 
-  static const Country NONE = Country._('');
+  const Country({
+    required this.code,
+  });
 
-  static const Country AFGHANISTAN = Country._('AF');
-  static const Country ALAND_ISLANDS = Country._('AX');
-  static const Country ALBANIA = Country._('AL');
-  static const Country ALGERIA = Country._('DZ');
-  static const Country AMERICAN_SAMOA = Country._('AS');
-  static const Country ANDORRA = Country._('AD');
-  static const Country ANGOLA = Country._('AO');
-  static const Country ANGUILLA = Country._('AI');
-  static const Country ANTARCTICA = Country._('AQ');
-  static const Country ANTIGUA_AND_BARBUDA = Country._('AG');
-  static const Country ARGENTINA = Country._('AR');
-  static const Country ARMENIA = Country._('AM');
-  static const Country ARUBA = Country._('AW');
-  static const Country AUSTRALIA = Country._('AU');
-  static const Country AUSTRIA = Country._('AT');
-  static const Country AZERBAIJAN = Country._('AZ');
-  static const Country BAHAMAS = Country._('BS');
-  static const Country BAHRAIN = Country._('BH');
-  static const Country BANGLADESH = Country._('BD');
-  static const Country BARBADOS = Country._('BB');
-  static const Country BELARUS = Country._('BY');
-  static const Country BELGIUM = Country._('BE');
-  static const Country BELIZE = Country._('BZ');
-  static const Country BENIN = Country._('BJ');
-  static const Country BERMUDA = Country._('BM');
-  static const Country BHUTAN = Country._('BT');
-  static const Country BOLIVIA = Country._('BO');
-  static const Country BONAIRE = Country._('BQ');
-  static const Country BOSNIA_AND_HERZEGOVINA = Country._('BA');
-  static const Country BOTSWANA = Country._('BW');
-  static const Country BOUVET_ISLAND = Country._('BV');
-  static const Country BRAZIL = Country._('BR');
-  static const Country BRITISH_INDIAN_OCEAN_TERRITORY = Country._('IO');
-  static const Country BRITISH_VIRGIN_ISLANDS = Country._('VG');
-  static const Country BRUNEI_DARUSSALAM = Country._('BN');
-  static const Country BULGARIA = Country._('BG');
-  static const Country BURKINA_FASO = Country._('BF');
-  static const Country BURUNDI = Country._('BI');
-  static const Country CABO_VERDE = Country._('CV');
-  static const Country CAMBODIA = Country._('KH');
-  static const Country CAMEROON = Country._('CM');
-  static const Country CANADA = Country._('CA');
-  static const Country CAYMAN_ISLANDS = Country._('KY');
-  static const Country CENTRAL_AFRICAN_REPUBLIC = Country._('CF');
-  static const Country CHAD = Country._('TD');
-  static const Country CHILE = Country._('CL');
-  static const Country CHINA = Country._('CN');
-  static const Country CHRISTMAS_ISLAND = Country._('CX');
-  static const Country COCOS_ISLANDS = Country._('CC');
-  static const Country COLOMBIA = Country._('CO');
-  static const Country COMOROS = Country._('KM');
-  static const Country CONGO = Country._('CG');
-  static const Country CONGO_DEMOCRATIC_REPUBLIC_OF = Country._('CD');
-  static const Country COOK_ISLANDS = Country._('CK');
-  static const Country COSTA_RICA = Country._('CR');
-  static const Country COTE_D_IVOIRE = Country._('CI');
-  static const Country CROATIA = Country._('HR');
-  static const Country CUBA = Country._('CU');
-  static const Country CURACAO = Country._('CW');
-  static const Country CYPRUS = Country._('CY');
-  static const Country CZECHIA = Country._('CZ');
-  static const Country DENMARK = Country._('DK');
-  static const Country DJIBOUTI = Country._('DJ');
-  static const Country DOMINICA = Country._('DM');
-  static const Country DOMINICAN_REPUBLIC = Country._('DO');
-  static const Country ECUADOR = Country._('EC');
-  static const Country EGYPT = Country._('EG');
-  static const Country EL_SALVADOR = Country._('SV');
-  static const Country EQUATORIAL_GUINEA = Country._('GQ');
-  static const Country ERITREA = Country._('ER');
-  static const Country ESTONIA = Country._('EE');
-  static const Country ETHIOPIA = Country._('ET');
-  static const Country FALKLAND_ISLANDS = Country._('FK');
-  static const Country FAROE_ISLANDS = Country._('FO');
-  static const Country FIJI = Country._('FJ');
-  static const Country FINLAND = Country._('FI');
-  static const Country FRANCE = Country._('FR');
-  static const Country FRENCH_GUIANA = Country._('GF');
-  static const Country FRENCH_POLYNESIA = Country._('PF');
-  static const Country FRENCH_SOUTHERN_TERRITORIES = Country._('TF');
-  static const Country GABON = Country._('GA');
-  static const Country GAMBIA = Country._('GM');
-  static const Country GEORGIA = Country._('GE');
-  static const Country GERMANY = Country._('DE');
-  static const Country GHANA = Country._('GH');
-  static const Country GIBRALTAR = Country._('GI');
-  static const Country GREECE = Country._('GR');
-  static const Country GREENLAND = Country._('GL');
-  static const Country GRENADA = Country._('GD');
-  static const Country GUADELOUPE = Country._('GP');
-  static const Country GUAM = Country._('GU');
-  static const Country GUATEMALA = Country._('GT');
-  static const Country GUERNSEY = Country._('GG');
-  static const Country GUINEA = Country._('GN');
-  static const Country GUINEA_BISSAU = Country._('GW');
-  static const Country GUYANA = Country._('GY');
-  static const Country HAITI = Country._('HT');
-  static const Country HEARD_ISLAND_AND_MCDONALD_ISLANDS = Country._('HM');
-  static const Country HONDURAS = Country._('HN');
-  static const Country HONG_KONG = Country._('HK');
-  static const Country HUNGARY = Country._('HU');
-  static const Country ICELAND = Country._('IS');
-  static const Country INDIA = Country._('IN');
-  static const Country INDONESIA = Country._('ID');
-  static const Country IRAN = Country._('IR');
-  static const Country IRAQ = Country._('IQ');
-  static const Country IRELAND = Country._('IE');
-  static const Country ISLE_OF_MAN = Country._('IM');
-  static const Country ISRAEL = Country._('IL');
-  static const Country ITALY = Country._('IT');
-  static const Country JAMAICA = Country._('JM');
-  static const Country JAPAN = Country._('JP');
-  static const Country JERSEY = Country._('JE');
-  static const Country JORDAN = Country._('JO');
-  static const Country KAZAKHSTAN = Country._('KZ');
-  static const Country KENYA = Country._('KE');
-  static const Country KIRIBATI = Country._('KI');
-  static const Country KUWAIT = Country._('KW');
-  static const Country KYRGYZSTAN = Country._('KG');
-  static const Country LAOS = Country._('LA');
-  static const Country LATVIA = Country._('LV');
-  static const Country LEBANON = Country._('LB');
-  static const Country LESOTHO = Country._('LS');
-  static const Country LIBERIA = Country._('LR');
-  static const Country LIBYA = Country._('LY');
-  static const Country LIECHTENSTEIN = Country._('LI');
-  static const Country LITHUANIA = Country._('LT');
-  static const Country LUXEMBOURG = Country._('LU');
-  static const Country MACAO = Country._('MO');
-  static const Country MACEDONIA = Country._('MK');
-  static const Country MADAGASCAR = Country._('MG');
-  static const Country MALAWI = Country._('MW');
-  static const Country MALAYSIA = Country._('MY');
-  static const Country MALDIVES = Country._('MV');
-  static const Country MALI = Country._('ML');
-  static const Country MALTA = Country._('MT');
-  static const Country MARSHALL_ISLANDS = Country._('MH');
-  static const Country MARTINIQUE = Country._('MQ');
-  static const Country MAURITANIA = Country._('MR');
-  static const Country MAURITIUS = Country._('MU');
-  static const Country MAYOTTE = Country._('YT');
-  static const Country MEXICO = Country._('MX');
-  static const Country MICRONESIA = Country._('FM');
-  static const Country MOLDOVA = Country._('MD');
-  static const Country MONACO = Country._('MC');
-  static const Country MONGOLIA = Country._('MN');
-  static const Country MONTENEGRO = Country._('ME');
-  static const Country MONTSERRAT = Country._('MS');
-  static const Country MOROCCO = Country._('MA');
-  static const Country MOZAMBIQUE = Country._('MZ');
-  static const Country MYANMAR = Country._('MM');
-  static const Country NAMIBIA = Country._('NA');
-  static const Country NAURU = Country._('NR');
-  static const Country NEPAL = Country._('NP');
-  static const Country NETHERLANDS = Country._('NL');
-  static const Country NEW_CALEDONIA = Country._('NC');
-  static const Country NEW_ZEALAND = Country._('NZ');
-  static const Country NICARAGUA = Country._('NI');
-  static const Country NIGER = Country._('NE');
-  static const Country NIGERIA = Country._('NG');
-  static const Country NIUE = Country._('NU');
-  static const Country NORFOLK_ISLAND = Country._('NF');
-  static const Country NORTH_KOREA = Country._('KP');
-  static const Country NORTHERN_MARIANA_ISLANDS = Country._('MP');
-  static const Country NORWAY = Country._('NO');
-  static const Country OMAN = Country._('OM');
-  static const Country PAKISTAN = Country._('PK');
-  static const Country PALAU = Country._('PW');
-  static const Country PALESTINE = Country._('PS');
-  static const Country PANAMA = Country._('PA');
-  static const Country PAPUA_NEW_GUINEA = Country._('PG');
-  static const Country PARAGUAY = Country._('PY');
-  static const Country PERU = Country._('PE');
-  static const Country PHILIPPINES = Country._('PH');
-  static const Country PITCAIRN = Country._('PN');
-  static const Country POLAND = Country._('PL');
-  static const Country PORTUGAL = Country._('PT');
-  static const Country PUERTO_RICO = Country._('PR');
-  static const Country QATAR = Country._('QA');
-  static const Country REUNION = Country._('RE');
-  static const Country ROMANIA = Country._('RO');
-  static const Country RUSSIAN_FEDERATION = Country._('RU');
-  static const Country RWANDA = Country._('RW');
-  static const Country SAINT_BARTHELEMY = Country._('BL');
-  static const Country SAINT_HELENA = Country._('SH');
-  static const Country SAINT_KITTS_AND_NEVIS = Country._('KN');
-  static const Country SAINT_LUCIA = Country._('LC');
-  static const Country SAINT_MARTIN = Country._('MF');
-  static const Country SAINT_PIERRE_AND_MIQUELON = Country._('PM');
-  static const Country SAINT_VINCENT_AND_THE_GRENADINES = Country._('VC');
-  static const Country SAMOA = Country._('WS');
-  static const Country SAN_MARINO = Country._('SM');
-  static const Country SAO_TOME_AND_PRINCIPE = Country._('ST');
-  static const Country SAUDI_ARABIA = Country._('SA');
-  static const Country SENEGAL = Country._('SN');
-  static const Country SERBIA = Country._('RS');
-  static const Country SEYCHELLES = Country._('SC');
-  static const Country SIERRA_LEONE = Country._('SL');
-  static const Country SINGAPORE = Country._('SG');
-  static const Country SINT_MAARTEN = Country._('SX');
-  static const Country SLOVAKIA = Country._('SK');
-  static const Country SLOVENIA = Country._('SI');
-  static const Country SOLOMON_ISLANDS = Country._('SB');
-  static const Country SOMALIA = Country._('SO');
-  static const Country SOUTH_AFRICA = Country._('ZA');
-  static const Country SOUTH_GEORGIA_AND_THE_SOUTH_SANDWICH_ISLANDS = Country._('GS');
-  static const Country SOUTH_KOREA = Country._('KR');
-  static const Country SOUTH_SUDAN = Country._('SS');
-  static const Country SPAIN = Country._('ES');
-  static const Country SRI_LANKA = Country._('LK');
-  static const Country SUDAN = Country._('SD');
-  static const Country SURINAME = Country._('SR');
-  static const Country SVALBARD_AND_JAN_MAYEN = Country._('SJ');
-  static const Country SWAZILAND = Country._('SZ');
-  static const Country SWEDEN = Country._('SE');
-  static const Country SWITZERLAND = Country._('CH');
-  static const Country SYRIAN_ARAB_REPUBLIC = Country._('SY');
-  static const Country TAIWAN = Country._('TW');
-  static const Country TAJIKISTAN = Country._('TJ');
-  static const Country TANZANIA = Country._('TZ');
-  static const Country THAILAND = Country._('TH');
-  static const Country TIMOR_LESTE = Country._('TL');
-  static const Country TOGO = Country._('TG');
-  static const Country TOKELAU = Country._('TK');
-  static const Country TONGA = Country._('TO');
-  static const Country TRINIDAD_AND_TOBAGO = Country._('TT');
-  static const Country TUNISIA = Country._('TN');
-  static const Country TURKEY = Country._('TR');
-  static const Country TURKMENISTAN = Country._('TM');
-  static const Country TURKS_AND_CAICOS_ISLANDS = Country._('TC');
-  static const Country TUVALU = Country._('TV');
-  static const Country UGANDA = Country._('UG');
-  static const Country UKRAINE = Country._('UA');
-  static const Country UNITED_ARAB_EMIRATES = Country._('AE');
-  static const Country UNITED_KINGDOM = Country._('GB');
-  static const Country UNITED_STATES = Country._('US');
-  static const Country UNITED_STATES_MINOR_OUTLYING_ISLANDS = Country._('UM');
-  static const Country URUGUAY = Country._('UY');
-  static const Country US_VIRGIN_ISLANDS = Country._('VI');
-  static const Country UZBEKISTAN = Country._('UZ');
-  static const Country VANUATU = Country._('VU');
-  static const Country VATICAN_CITY = Country._('VA');
-  static const Country VENEZUELA = Country._('VE');
-  static const Country VIET_NAM = Country._('VN');
-  static const Country WALLIS_AND_FUTUNA = Country._('WF');
-  static const Country WESTERN_SAHARA = Country._('EH');
-  static const Country YEMEN = Country._('YE');
-  static const Country ZAMBIA = Country._('ZM');
-  static const Country ZIMBABWE = Country._('ZW');
-
-  const Country._(String countryCode) : _countryCode = countryCode;
-
-  /// Returns the string value of the Country code.
-  String get countryCode => _countryCode;
+  final String code;
 }