+++ lib/src/search/search.dart
@@ -2,13 +2,8 @@
 // MIT license that can be found in the LICENSE file.
 
 import 'package:podcast_search/podcast_search.dart';
-import 'package:podcast_search/src/model/attribute.dart';
-import 'package:podcast_search/src/model/country.dart';
-import 'package:podcast_search/src/model/language.dart';
-import 'package:podcast_search/src/model/search_result.dart';
 import 'package:podcast_search/src/search/itunes_search.dart';
 import 'package:podcast_search/src/search/podcast_index_search.dart';
-import 'package:podcast_search/src/search/providers/providers.dart';
 
 /// This class handles the searching. Taking the base URL we build any parameters
 /// that have been added before making a call to iTunes. The results are unpacked
@@ -18,20 +13,20 @@ class Search {
   String _term = '';
 
   /// If this property is not-null search results will be limited to this country
-  Country _country = Country.NONE;
+  Country _country = Country.none;
 
   /// If this property is not-null search results will be limited to this genre
   String _genre = '';
 
   /// By default, searches will be performed against keyword(s) in [_term].
   /// Set this property to search against a different attribute.
-  Attribute _attribute = Attribute.NONE;
+  Attribute _attribute = Attribute.none;
 
   /// Limit the number of results to [_limit]. If zero no limit will be applied
   int _limit = 0;
 
   /// If non-null, the results will be limited to the language specified.
-  Language _language = Language.NONE;
+  Language _language = Language.none;
 
   /// Set to true to disable the explicit filter.
   bool _explicit = false;
@@ -57,9 +52,9 @@ class Search {
   /// value to search by a different attribute such as Author, genre etc.
   Future<SearchResult> search(
     String term, {
-    Country country = Country.NONE,
-    Attribute attribute = Attribute.NONE,
-    Language language = Language.NONE,
+    Country country = Country.none,
+    Attribute attribute = Attribute.none,
+    Language language = Language.none,
     int limit = 0,
     int version = 0,
     bool explicit = false,
@@ -97,7 +92,7 @@ class Search {
 
   /// Fetches the list of top podcasts
   /// Optionally takes a [limit] and [Country] filter. Defaults to
-  /// limit of 20 and the UK.
+  /// limit of 20 and no specified country.
   ///
   /// The charts is returned as a 'feed'. In order to be compatible with
   /// [SearchResult] we need to parse this feed and fetch the underlying
@@ -105,7 +100,7 @@ class Search {
   /// the infrequent update of the chart feed it is recommended that clients
   /// cache the results.
   Future<SearchResult> charts({
-    Country country = Country.UNITED_KINGDOM,
+    Country country = Country.none,
     int limit = 20,
     bool explicit = false,
     String genre = '',